feat(agentics): refactor components to agenerate, amap, areduce pattern
- Renamed semantic_map to amap_component
- Renamed semantic_aggregator to areduce_component
- Added new agenerate_component for parallel generation
- Updated Agentics logo/icon
- Updated component icons and documentation
- Removed synthetic_data_generator (replaced by agenerate)
- Updated component_index.json with new component structure
