+++ src/lfx/src/lfx/components/agentics/constants.py
@@ -50,7 +50,7 @@ TRANSDUCTION_TYPES = [TRANSDUCTION_AMAP, TRANSDUCTION_AREDUCE, TRANSDUCTION_GENE
 
 # Error messages for user feedback
 ERROR_AGENTICS_NOT_INSTALLED = (
-    "Agentics-py is not installed. Please install it with `uv pip install agentics-py==0.3.1`."
+    "Agentics-py is not installed. Please install it with `uv pip install agentics-py`."
 )
 ERROR_API_KEY_REQUIRED = "{provider} API key is required. Please configure it in your settings or provide it directly."
 ERROR_UNSUPPORTED_PROVIDER = (
